Simplify message_queue_message_alloc
We were speculatively reading the freelist head hoping to avoid locking the
spinlock if there's no free memory to reuse anyway. However, this could lead
to someone else sniping the last free message, causing us to have to retry.
So instead we'll go ahead and obtain the lock, since it only wastes time on
the slow path through malloc anyway.