Import memory allocator
This allocator avoids using spinlocks by using a ring buffer of pointers to
store the freelist. Allocating involves decrementing the free block count,
backing out if it drops below 0, then incrementing the next free block
position, waiting for the pointer at the old position to become valid, then
returning that pointer. Freeing is a similar process, ensuring that the slot
we're about to write to is NULL to make sure any in-progress allocations
have had a chance to finish before overwriting their pointers.